Thin injection.js + messageRouter
Thin out injection.js, to only contain entry for bootstraps + wires runtime message router.

messageRouter.js replaces the if/else chain with a dispatch table. Action keys preserved exactly (including the historical `'display: encouragement'` and `'display: rewardOverlay'` whitespace).
